Saving: put back tip_distance
In a previous change: saved area location was changed from depth from insertion (arbitrary relative to channel) to distance from tip (absolute measurement).

This change was lost over the course of previous updates, so this functionality was restored here.

Now areas location is saved as probe_areas{#probe}.tip_distance, which is the distance of the area borders from the shank tip along the entire trajectory.